+++ tests/test_metrics_mql_query.py
@@ -551,6 +551,45 @@ metrics_query_timeseries_to_mql_tests = [
         },
         id="test_crazy_characters",
     ),
+    pytest.param(
+        MetricsQuery(
+            query="sum(transaction.duration){status_code:500} by transaction",
+            start=NOW,
+            end=NOW + timedelta(days=14),
+            rollup=Rollup(interval=3600, totals=None, granularity=3600),
+            scope=MetricsScope(
+                org_ids=[1], project_ids=[11], use_case_id="transactions"
+            ),
+            limit=Limit(100),
+            offset=Offset(5),
+            indexer_mappings={},
+        ),
+        {
+            "mql": 'sum(transaction.duration){status_code:"500"} by (transaction)',
+            "mql_context": {
+                "entity": {
+                    "transaction.duration": None,
+                },
+                "start": "2023-01-02T03:04:05+00:00",
+                "end": "2023-01-16T03:04:05+00:00",
+                "rollup": {
+                    "orderby": None,
+                    "granularity": 3600,
+                    "interval": 3600,
+                    "with_totals": None,
+                },
+                "scope": {
+                    "org_ids": [1],
+                    "project_ids": [11],
+                    "use_case_id": "transactions",
+                },
+                "limit": 100,
+                "offset": 5,
+                "indexer_mappings": {},
+            },
+        },
+        id="test_passing_string_directly",
+    ),
 ]